requirements.txt
streamlit
pandas
openpyxl
google-cloud-vision
PyMuPDF
pillow

